Jira/requirement.txt
asn1crypto==0.24.0
cachetools==2.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
enum34==1.1.6
google-auth==1.5.1
idna==2.7
ipaddress==1.0.22
jira==2.0.0
oauthlib==2.1.0
pbr==5.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
requests==2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.11.0
urllib3==1.24
